--- dubbo-reactive/src/main/java/org/apache/dubbo/rpc/protocol/tri/reactive/ClientTripleReactorSubscriber.java
@@ -1,34 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.dubbo.rpc.protocol.tri.reactive;
-
-import org.apache.dubbo.rpc.protocol.tri.observer.ClientCallToObserverAdapter;
-
-/**
- * The subscriber in client to subscribe user publisher and is subscribed by ClientStreamObserver.
- */
-public class ClientTripleReactorSubscriber<T> extends AbstractTripleReactorSubscriber<T> {
-
-    @Override
-    public void cancel() {
-        if (!isCanceled()) {
-            super.cancel();
-            ((ClientCallToObserverAdapter<T>) downstream).cancel(new Exception("Cancelled"));
-        }
-    }
-}
