--- dubbo-reactive/src/main/java/org/apache/dubbo/rpc/protocol/tri/reactive/ServerTripleReactorPublisher.java
@@ -1,33 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.dubbo.rpc.protocol.tri.reactive;
-
-import org.apache.dubbo.rpc.protocol.tri.observer.CallStreamObserver;
-
-/**
- * Used in ManyToOne and ManyToMany in server. <br>
- * It is a Publisher for user subscriber to subscribe. <br>
- * It is a StreamObserver for requestStream. <br>
- * It is a Subscription for user subscriber to request and pass request to responseStream.
- */
-public class ServerTripleReactorPublisher<T> extends AbstractTripleReactorPublisher<T> {
-
-    public ServerTripleReactorPublisher(CallStreamObserver<?> callStreamObserver) {
-        super.onSubscribe(callStreamObserver);
-    }
-}
